Move JsonPayloadConfig to the Bootloader\Http
