+++ src/Framework/Http/Middleware/JsonPayloadMiddleware.php
@@ -15,7 +15,7 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Spiral\Config\JsonPayloadConfig;
+use Spiral\Bootloader\Http\JsonPayloadConfig;
 use Spiral\Http\Exception\ClientException;
 
 /**
